만료법인 삭제 에러 수정 (#627)
* 스케쥴러, 이메일 템플릿 추가

* 템플릿수정, cron식 변경, sql조건문변경

* 원복

* 변수명변경, 메일수신자변경,

* prod환경에서만 스케듈 돌도록 수정

* 코멘트 반영

* pr 코멘트내역 반영

* png 경로 변경

* isInited -> isReset 변경

* 로직변경

* 누락추가

* 누락추가

* 누락추가

* 누락추가

* 운영1에서만 검증하도록 로직변경 & User정보 초기화로직 테스트 및 수정

* 오타수정

* 삭제오류 수정

+++ gm-api-user/src/main/java/com/nomadconnection/dapp/api/service/UserService.java
@@ -147,24 +147,26 @@ public class UserService {
 	@Transactional(rollbackFor = Exception.class)
 	public void initUserInfo(Long idxUser) {
 		User user = getUser(idxUser);
-		Long idxCorp = user.corp().idx();
-        List<Long> listIdxCardInfo = repoCardIssuanceInfo.findAllIdxByUserIdx(idxUser);
-        repoCeoInfo.deleteAllByCardIssuanceInfoIdx(listIdxCardInfo);
-        repoManager.deleteAllByCardIssuanceInfoIdx(listIdxCardInfo);
-        repoStockholderFile.deleteAllByCardIssuanceInfoIdx(listIdxCardInfo);
+        if (!ObjectUtils.isEmpty(user.corp())) {
+            Long idxCorp = user.corp().idx();
+            List<Long> listIdxCardInfo = repoCardIssuanceInfo.findAllIdxByUserIdx(idxUser);
+            repoCeoInfo.deleteAllByCardIssuanceInfoIdx(listIdxCardInfo);
+            repoManager.deleteAllByCardIssuanceInfoIdx(listIdxCardInfo);
+            repoStockholderFile.deleteAllByCardIssuanceInfoIdx(listIdxCardInfo);
+            fullTextService.deleteAllShinhanFulltext(idxCorp);
+            fullTextService.deleteAllLotteFulltext(idxCorp);
+            repoRisk.deleteByCorpIdx(idxCorp);
+            // Todo 추후 corp테이블의 idxRiskConfig 값 & AdminService 수정시 반영
+            repoRiskConfig.delete(user.corp().riskConfig());
+            repoRiskConfig.flush();
+            repoCorp.delete(user.corp());
+            repoCorp.flush();
+            user.corp().riskConfig(null);
+        }
         repoCardIssuanceInfo.deleteAllByUserIdx(idxUser);
         repoConnectdMng.deleteAllByUserIdx(idxUser);
         repoConsentMapping.deleteAllByUserIdx(idxUser);
         issuanceProgressRepository.deleteAllByUserIdx(idxUser);
-        fullTextService.deleteAllShinhanFulltext(idxCorp);
-        fullTextService.deleteAllLotteFulltext(idxCorp);
-        repoRisk.deleteByCorpIdx(idxCorp);
-        // Todo 추후 corp테이블의 idxRiskConfig 값 & AdminService 수정시 반영
-		repoRiskConfig.delete(user.corp().riskConfig());
-		repoRiskConfig.flush();
-		repoCorp.delete(user.corp());
-		repoCorp.flush();
-		user.corp().riskConfig(null);
 		user.corp(null);
 		user.cardCompany(null);
 		user.isReset(true);
